Change ID type to long

+++ src/main/java/com/climateconfort/data_reporter/cassandra/domain/gela/Gela.java
@@ -12,8 +12,8 @@ import lombok.Setter;
 @Entity
 public class Gela {
     @PartitionKey
-    int id;
+    long id;
 
     @CqlName("eraikina_id")
-    int eraikinaId;
+    long eraikinaId;
 }
